pyproject.toml
[project]
name = "zigi-amazon-mcp"
version = "0.0.1"
description = "This is a prototype MCP server to connect to the Amazon Seller Central API."
authors = [{ name = "GMCLG", email = "pelmog@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastmcp>=2.5.1",
    "mcp>=1.0.0",
    "boto3>=1.34.0",
    "requests>=2.31.0",
    "requests-aws4auth>=1.2.3",
    "python-dotenv>=1.0.0",
    "jsonquerylang>=0.1.0",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://Pelmoggian.github.io/zigi-amazon-mcp/"
Repository = "https://github.com/Pelmoggian/zigi-amazon-mcp"
Documentation = "https://Pelmoggian.github.io/zigi-amazon-mcp/"

[project.scripts]
zigi-amazon-mcp = "zigi_amazon_mcp.server:main"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.23.0",
    "mypy>=0.991",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "pytest-asyncio>=1.0.0",
    "types-requests>=2.32.0.20250515",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zigi_amazon_mcp"]

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S108"]
"src/zigi_amazon_mcp/server.py" = ["TRY003", "TRY300", "TRY301", "S307", "C901"]
"src/zigi_amazon_mcp/inventory_sample.py" = ["C901"]
"OldAmazonCode/*" = ["ALL"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
